Implemented, tested, and cleansed.
