+++ Makefile
@@ -70,19 +70,20 @@ resetpy: ALWAYS
 test: ALWAYS
 	pip install -r requirements.txt
 	pip install -e .
-	pytest -v ./test/test-baggage.py
-	pytest -v ./test/test-boardingpass.py
-	pytest -v ./test/test-customforms.py
-	pytest -v ./test/test-expenses.py
-	pytest -v ./test/test-passportscanner.py
-	pytest -v ./test/test-trees.py
 	pytest -v ./test/test-util.py
-	pytest -v ./test/test-validate.py
-	pytest -v ./test/test-gamecomp.py
-	pytest -v ./test/test-xmascypher.py
-	pytest -v ./test/test-adapters.py
-	pytest -v ./test/test-ship.py
-	pytest -v ./test/test-bus.py
+	pytest -v ./test/test_day_01_expenses.py
+	pytest -v ./test/test_day_02_validate.py
+	pytest -v ./test/test_day_03_trees.py
+	pytest -v ./test/test_day_04_passportscanner.py
+	pytest -v ./test/test_day_05_boardingpass.py
+	pytest -v ./test/test_day_06_customforms.py
+	pytest -v ./test/test_day_07_baggage.py
+	pytest -v ./test/test_day_08_gamecomp.py
+	pytest -v ./test/test_day_09_xmascypher.py
+	pytest -v ./test/test_day_10_adapters.py
+	pytest -v ./test/test_day_11_seats.py
+	pytest -v ./test/test_day_12_ship.py
+	pytest -v ./test/test_day_13_bus.py
 	pip uninstall -y $(MODULE)==$(VERSION) || true
 	rm -Rfv $$(find $(MODULE) | awk '/__pycache__$$/')
 	rm -Rfv $$(find ./aoc | awk '/__pycache__$$/')
