+++ aoc/day_01_expenses.py
