+++ aoc/day_02_validate.py
