+++ aoc/day_03_trees.py
