+++ aoc/day_05_boardingpass-mini.py
