+++ aoc/day_06_customforms.py
