+++ aoc/day_07_baggage.py
