+++ aoc/day_10_adapters.py
