+++ aoc/day_11_seats.py
