+++ aoc/day_12_ship.py
